Add React Storybook, for development
About to start work on enter/exit animations, and I'm sick of not
having a good dev experience for this project.

React Storyboard is meant for managing the design and UI of components
in a large application, but it suits my needs just fine. Gives me a
hot-reloaded, transpiled server to work in, without significantly
making the repo more complex (a config file, and a `stories` dir).

Setup is ready; I have a list that randomly selects 4 out of a large
list, and so enter/exit animations can be triggered with a button.